requirements.txt
bs4
vaderSentiment
schedule
websocket-client
lxml
websockets
aiokafka
kafka-python
pyspark